requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.1
gunicorn==21.2.0